rename gguf folder to quantization
